requirements.txt
amqp==5.1.1
arrow==1.2.3
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
binaryornot==0.4.4
celery==5.2.7
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cookiecutter==2.1.1
cron-descriptor==1.2.35
Django==4.2
django-browser-reload==1.7.0
django-celery-beat==2.5.0
django-tailwind==3.5.0
django-timezone-field==5.0
django-versatileimagefield==3.0
djangorestframework==3.14.0
flake8==6.0.0
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
kombu==5.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-magic==0.4.27
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
redis==4.5.4
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
